EndState must be a bool, not None
In class `StateNextEnd`, the `EndState` value is checked to ensure it is a boolean. In the subclasses however, the default value for `EndState` is `None`, which means if `EndState` is not specified, it would fail as the default value of `None` is not an instance of the bool type.

This PR changes the subclass `__init__` methods so the default value for `EndState` is `False` instead. This matches up with the doc string.